requirements.txt
streamlit 
langchain
langchain-community
langchain-google-genai
python-dotenv